Create The Minimal API for the System
Created the Minimal API for the System. The blazor app is not yet calling the API.

In the database we make the address optional. Restrict the deletion of the merchant.

In the future we need to implement a way to clean up the database since In this commit we change the ServiceBooking Table Delete behavior to NoAction since the database had a conflict.

We also inject the properly the services and database. In the future we might need to create a static function to inject all the services so that the program would be readable.